fix: 프론트 훅 lint 구조 정리
