+++ frontend/src/shared/i18n/index.ts
@@ -1,0 +1,3 @@
+export { I18nProvider } from "./provider";
+export { useI18n } from "./use-i18n";
+export { SUPPORTED_LOCALES, type Locale } from "./resources";
