+++ frontend/src/shared/i18n/use-i18n.ts
@@ -1,0 +1,12 @@
+import { useContext } from "react";
+import { I18nContext } from "./context";
+
+export function useI18n() {
+  const context = useContext(I18nContext);
+
+  if (!context) {
+    throw new Error("useI18n must be used within an I18nProvider.");
+  }
+
+  return context;
+}
